pyproject.toml
[tool.poetry]
name = "query-vertica"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
vertica-python = "^1.3.8"
pyyaml = "^6.0.1"
pytest = "^8.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
